Primeira versão do algoritmo genético para as paradas de ônibus de Porto Alegre.

+++ Ponto.cpp
@@ -1,57 +1,48 @@
-// 
-//  Ponto.cpp
-//  OpenGLTest
-//
-//  Created by Márcio Sarroglia Pinho on 18/08/20.
-//  Copyright © 2020 Márcio Sarroglia Pinho. All rights reserved.
-//
+#include "Ponto.hpp"
 
-#include "Ponto.h"
-Ponto::Ponto ()
-{
+Ponto::Ponto(){
     x=y=z=0;
 }
-Ponto::Ponto(float x, float y, float z)
-{
+
+Ponto::Ponto(double x, double y, double z){
     this->x = x;
     this->y = y;
     this->z = z;
 }
-void Ponto::set(float x, float y, float z)
-{
+
+void Ponto::set(double x, double y, double z){
     this->x = x;
     this->y = y;
     this->z = z;
 }
-void Ponto::imprime() {
+
+void Ponto::imprime(){
     cout << "(" << x << ", " << y << ", " << z <<")" << flush;
 }
-void Ponto::imprime(char const *msg)
-{
+
+void Ponto::imprime(char const *msg){
     cout << msg;
     imprime();
 }
-void Ponto::imprime(char const *msgAntes, char const *msgDepois)
-{
+
+void Ponto::imprime(char const *msgAntes, char const *msgDepois){
     imprime(msgAntes);
     cout << msgDepois;
 }
-void Ponto::multiplica(double x, double y, double z)
-{
+
+void Ponto::multiplica(double x, double y, double z){
     this->x *= x;
     this->y *= y;
     this->z *= z;
 }
 
-void Ponto::soma(double x, double y, double z)
-{
+void Ponto::soma(double x, double y, double z){
     this->x += x;
     this->y += y;
     this->z += z;
 }
 
-void Ponto::rotacionaZ(float angulo)
-{
+void Ponto::rotacionaZ(float angulo){
     float xr, yr;
     //cout << "Angulo: " << angulo << " ";
     double anguloRad = angulo * 3.14159265359/180.0;
@@ -63,8 +54,7 @@ void Ponto::rotacionaZ(float angulo)
 }
 
 
-void Ponto::rotacionaY(float angulo)
-{
+void Ponto::rotacionaY(float angulo){
     float xr, zr;
     double anguloRad = angulo* 3.14159265359/180.0;
     xr =  x*cos(anguloRad) + z*sin(anguloRad);
@@ -73,8 +63,7 @@ void Ponto::rotacionaY(float angulo)
     z = zr;
 }
 
-void Ponto::rotacionaX(float angulo)
-{
+void Ponto::rotacionaX(float angulo){
     float yr, zr;
     double anguloRad = angulo* 3.14159265359/180.0;
     yr =  y*cos(anguloRad) - z*sin(anguloRad);
@@ -82,13 +71,11 @@ void Ponto::rotacionaX(float angulo)
     y = yr;
     z = zr;
 }
-double Ponto::modulo()
-{
+double Ponto::modulo(){
     return sqrt(x*x+y*y+z*z);
 }
 
-void Ponto::versor()
-{
+void Ponto::versor(){
     float m = modulo();
     x /= m;
     y /= m;
@@ -96,8 +83,7 @@ void Ponto::versor()
 }
 
 
-Ponto ObtemMaximo (Ponto P1, Ponto P2)
-{
+Ponto ObtemMaximo(Ponto P1, Ponto P2){
     Ponto Max;
     
     Max.x = (P2.x > P1.x) ? P2.x : P1.x;
@@ -105,8 +91,8 @@ Ponto ObtemMaximo (Ponto P1, Ponto P2)
     Max.z = (P2.z > P1.x) ? P2.z : P1.z;
     return Max;
 }
-Ponto ObtemMinimo (Ponto P1, Ponto P2)
-{
+
+Ponto ObtemMinimo(Ponto P1, Ponto P2){
     Ponto Min;
     
     Min.x = (P2.x < P1.x) ? P2.x : P1.x;
@@ -114,16 +100,15 @@ Ponto ObtemMinimo (Ponto P1, Ponto P2)
     Min.z = (P2.z < P1.x) ? P2.z : P1.z;
     return Min;
 }
-bool operator==(Ponto P1, Ponto P2)
-{
+
+bool operator==(Ponto P1, Ponto P2){
     if (P1.x != P2.x) return false;
     if (P1.y != P2.y) return false;
     if (P1.z != P2.z) return false;
     return true;
-
 }
-Ponto operator+(Ponto P1, Ponto P2)
-{
+
+Ponto operator+(Ponto P1, Ponto P2){
     Ponto temp;
     temp = P1;
     temp.x += P2.x;
@@ -132,8 +117,7 @@ Ponto operator+(Ponto P1, Ponto P2)
     return temp;
 }
 
-Ponto operator- (Ponto P1, Ponto P2)
-{
+Ponto operator-(Ponto P1, Ponto P2){
     Ponto temp;
     temp = P1;
     temp.x -= P2.x;
@@ -141,8 +125,8 @@ Ponto operator- (Ponto P1, Ponto P2)
     temp.z -= P2.z;
     return temp;
 }
-Ponto operator* (Ponto P1, float k)
-{
+
+Ponto operator*(Ponto P1, double k){
     Ponto temp;
     temp.x = P1.x * k;
     temp.y = P1.y * k;
@@ -150,26 +134,24 @@ Ponto operator* (Ponto P1, float k)
     return temp;
 }
 
-Ponto operator-(Ponto P1)
-{
+Ponto operator-(Ponto P1){
     return P1 * -1;
 }
 // **********************************************************************
 //    Calcula o produto escalar entre os vetores V1 e V2
 // **********************************************************************
-double ProdEscalar(Ponto v1, Ponto v2)
-{
+double ProdEscalar(Ponto v1, Ponto v2){
     return v1.x*v2.x + v1.y*v2.y+ v1.z*v2.z;
 }
 // **********************************************************************
 //    Calcula o produto vetorial entre os vetores V1 e V2
 // **********************************************************************
-void ProdVetorial (Ponto v1, Ponto v2, Ponto &vresult)
-{
+void ProdVetorial (Ponto v1, Ponto v2, Ponto &vresult){
     vresult.x = v1.y * v2.z - (v1.z * v2.y);
     vresult.y = v1.z * v2.x - (v1.x * v2.z);
     vresult.z = v1.x * v2.y - (v1.y * v2.x);
 }
+
 long int ContadorInt=0;
 /* ********************************************************************** */
 /*                                                                        */
@@ -184,8 +166,7 @@ long int ContadorInt=0;
 /* t: valor do parâmetro no ponto de interseção (sobre a reta MN)         */
 /*                                                                        */
 /* ********************************************************************** */
-int intersec2d(Ponto k, Ponto l, Ponto m, Ponto n, double &s, double &t)
-{
+int intersec2d(Ponto k, Ponto l, Ponto m, Ponto n, double &s, double &t){
     double det;
 
     det = (n.x - m.x) * (l.y - k.y)  -  (n.y - m.y) * (l.x - k.x);
@@ -201,8 +182,7 @@ int intersec2d(Ponto k, Ponto l, Ponto m, Ponto n, double &s, double &t)
 // **********************************************************************
 //
 // **********************************************************************
-bool HaInterseccao(Ponto k, Ponto l, Ponto m, Ponto n)
-{
+bool HaInterseccao(Ponto k, Ponto l, Ponto m, Ponto n){
     int ret;
     double s,t;
     
@@ -217,15 +197,13 @@ bool HaInterseccao(Ponto k, Ponto l, Ponto m, Ponto n)
 // **********************************************************************
 //
 // **********************************************************************
-void resetContadorInt()
-{
+void resetContadorInt(){
     ContadorInt = 0;
 }
 // **********************************************************************
 //
 // **********************************************************************
-long int getContadorInt()
-{
+long int getContadorInt(){
     return ContadorInt;
 }
 
@@ -233,8 +211,7 @@ long int getContadorInt()
 // **********************************************************************
 //
 // **********************************************************************
-double calculaDistancia(Ponto P, Ponto Q)
-{
+double calculaDistancia(Ponto P, Ponto Q){
     float dx, dy, dz;
     
     dx = P.x - Q.x;
@@ -247,8 +224,7 @@ double calculaDistancia(Ponto P, Ponto Q)
 // **********************************************************************
 // int lado(Ponto P1, Ponto P2, Ponto A)
 // **********************************************************************
-int lado(Ponto P1, Ponto P2, Ponto A)
-{
+int lado(Ponto P1, Ponto P2, Ponto A){
     Ponto V1, V2, V3;
     V1 = P2-P1;
     V2 = A-P1;
